+++ src/packages/piral-preact/src/mount.ts
@@ -1,0 +1,5 @@
+import { render, ComponentType, createElement } from 'preact';
+
+export function mount<T extends Object>(el: HTMLElement, root: ComponentType<T>, props: T, ctx: any) {
+  render(createElement(root, props), el);
+}