+++ src/samples/sample-piral-core/src/pilets/index.ts
@@ -9,6 +9,7 @@ import { NgPilet } from './ngPilet';
 import { HyperappPilet } from './hyperappPilet';
 import { ContainerPilet } from './container';
 import { InfernoPilet } from './infernoPilet';
+import { PreactPilet } from './preactPilet';
 
 /**
  * Normally all these pilets would come from some API and
@@ -31,4 +32,5 @@ export const availablePilets = [
   HyperappPilet,
   ContainerPilet,
   InfernoPilet,
+  PreactPilet,
 ];